+++ spelling/docs/source/history.rst
@@ -2,6 +2,18 @@
  Release History
 =================
 
+trunk
+
+  - Add an option treat the names of packages on PyPI as spelled
+    properly.
+  - Add an option to treat CamelCase names as spelled properly.
+  - Add an option to treat acronyms as spelled properly.
+  - Add an option to treat Python built-ins as spelled properly.
+  - Add an option to treat names that can be found as modules as
+    spelled properly.
+  - Add an option to let the user provide a list of other filter
+    classes for the tokenizer.
+
 1.0
 
   - Re-implement using just a Builder, without a separate visitor